quest schemas are now better organized
