+++ src/hooks/use-quest.tsx
@@ -1,8 +1,8 @@
 import useSWR from "swr"
-import {ApiSchema} from "@/app/(admin)/admin/quests/editor/_types/api_schema";
+import {QuestSchema} from "@/app/(admin)/admin/quests/editor/_types/api_schema";
 
 
-export function useQuest(id: string | undefined): {quest: ApiSchema | null, isError: any, isLoading: boolean} {
+export function useQuest(id: string | undefined): {quest: QuestSchema | null, isError: any, isLoading: boolean} {
     const fetcher = (url: string) => fetch(url)
         .then((res) => res.json());
